finalizado modulo de adocao

+++ backend-buscapet/src/modules/adoptions/services/DeleteAdoptionService.ts
@@ -4,16 +4,27 @@ import AppError from '@shared/errors/AppError';
 
 interface IRequest {
   id: string;
-  user_id: string;
+  adopter_id: string;
+  isOng: boolean;
 }
 
 class DeleteAdoptionService {
-  public async execute({ id, user_id }: IRequest): Promise<void> {
+  public async execute({ id, adopter_id, isOng }: IRequest): Promise<void> {
     const adoptionsRepository = getCustomRepository(AdoptionsRepository);
 
-    const adoption = await adoptionsRepository.findById(id, user_id);
+    if (isOng) {
+      throw new AppError('JWT Token inválido');
+    }
+
+    const adoption = await adoptionsRepository.findById(id, adopter_id, isOng);
     if (!adoption) {
       throw new AppError('Adoção não encontrada!');
+    }
+
+    if (adoption.status !== 'Solicitada') {
+      throw new AppError(
+        'Adoção aprovada ou reprovada não pode ser cancelada!',
+      );
     }
 
     await adoptionsRepository.remove(adoption);
